Correção completa
Alterado Lumen para Quintana, corrigido a correção/verificação ortográfica, e inserido arquivo de configuração para trocar de modelos
