+++ frontend/src/components/modalDetalhesTema.tsx
@@ -1,20 +1,20 @@
 import { Modal, Input, Button, message } from 'antd';
 import { useState } from 'react';
-import { Tema } from '@/pages/lumen/home';
+import { Tema } from '@/pages/quintana/home';
 import { useAuth } from '@/context';
-import {API_URL} from "@/config/config";
+import { API_URL } from "@/config/config";
 
 interface TemaDetalhes {
     open: boolean;
     onCancel: () => void;
-    tema: Tema | null; 
+    tema: Tema | null;
     onTemaEditado: (temaEditado: Tema) => void;
 }
 
 const ModalDetalhesTema: React.FC<TemaDetalhes> = ({ open, onCancel, tema, onTemaEditado }) => {
     const [temaEditado, setTemaEditado] = useState<string>('');
     const [descricaoEditada, setDescricaoEditada] = useState<string>('');
-    const { tipoUsuario } = useAuth(); 
+    const { tipoUsuario } = useAuth();
 
     const handleEditarTema = async () => {
         try {
@@ -24,11 +24,11 @@ const ModalDetalhesTema: React.FC<TemaDetalhes> = ({ open, onCancel, tema, onTem
                     headers: {
                         'Content-Type': 'application/json'
                     },
-                    body: JSON.stringify({ 
-                        tema: temaEditado !== '' ? temaEditado : tema.tema, 
+                    body: JSON.stringify({
+                        tema: temaEditado !== '' ? temaEditado : tema.tema,
                         descricao: descricaoEditada !== '' ? descricaoEditada : tema.descricao,
                         nome_professor: tema.nome_professor
-                     })
+                    })
                 });
                 if (response.ok) {
                     message.success('Tema atualizado com sucesso!');
@@ -48,7 +48,7 @@ const ModalDetalhesTema: React.FC<TemaDetalhes> = ({ open, onCancel, tema, onTem
 
     return (
         <Modal
-            title= {tipoUsuario === 'aluno' ? 'Detalhes do Tema' : 'Editar Tema'}
+            title={tipoUsuario === 'aluno' ? 'Detalhes do Tema' : 'Editar Tema'}
             open={open}
             onCancel={onCancel}
             footer={null}
@@ -65,12 +65,12 @@ const ModalDetalhesTema: React.FC<TemaDetalhes> = ({ open, onCancel, tema, onTem
                 </div>
             ) : tema && (
                 <div>
-                    <label style={{marginBottom: '10px'}}><b>Professor</b>:</label>
-                    <Input style={{marginBottom: '10px'}} value={tema.nome_professor} disabled />
-                    <label style={{marginBottom: '10px'}}><b>Tema</b>:</label>
-                    <Input style={{marginBottom: '10px'}} defaultValue={tema.tema} onChange={(e) => setTemaEditado(e.target.value)} />
-                    <label style={{marginBottom: '10px'}}><b>Descrição</b>:</label>
-                    <Input.TextArea style={{marginBottom: '10px'}} defaultValue={tema.descricao} onChange={(e) => setDescricaoEditada(e.target.value)} />
+                    <label style={{ marginBottom: '10px' }}><b>Professor</b>:</label>
+                    <Input style={{ marginBottom: '10px' }} value={tema.nome_professor} disabled />
+                    <label style={{ marginBottom: '10px' }}><b>Tema</b>:</label>
+                    <Input style={{ marginBottom: '10px' }} defaultValue={tema.tema} onChange={(e) => setTemaEditado(e.target.value)} />
+                    <label style={{ marginBottom: '10px' }}><b>Descrição</b>:</label>
+                    <Input.TextArea style={{ marginBottom: '10px' }} defaultValue={tema.descricao} onChange={(e) => setDescricaoEditada(e.target.value)} />
                     <Button onClick={handleEditarTema}>Editar</Button>
                 </div>
             )}
