+++ frontend/src/pages/quintana/competencias.tsx
